fix: align tests and note-expression value cast with pinned VST3 SDK
- Cast NoteExpressionValueDescription::associatedParameterId as int32_t
  so the sentinel -1 is preserved (instead of surfacing as 4294967295).
- Update saveState tests to read the gain float at offset 9 (after the
  9-byte 'NST3' envelope header) rather than at offset 0.
- Replace assert.throws(fn, /VST3_/) regex matchers with code-based
  validators (err.code === 'VST3_*') across the test suite, since the
  regex matches the message, not the .code property.
- Treat numSamples=0 as a valid parameter-flush block (doesNotThrow).
- Accept boolean return from setKnobMode on the SingleComponentEffect
  fixture (implements IEditController2 via the base class).
- Drop the stale `info.type` assertion on UnitInfo (pinned SDK's
  UnitInfo has no `type` field).

All 150 tests now pass (147 pass, 3 skipped, 0 fail) locally.

+++ src/plugin_instance.cc
@@ -2003,7 +2003,8 @@ Napi::Value PluginInstance::GetNoteExpressionInfo(const Napi::CallbackInfo& info
         o.Set("shortTitle", Napi::String::New(env, string128ToUtf8(ne.shortTitle)));
         o.Set("unitId", Napi::Number::New(env, static_cast<double>(ne.unitId)));
         o.Set("associatedParameterId",
-              Napi::Number::New(env, static_cast<double>(ne.associatedParameterId)));
+              Napi::Number::New(env, static_cast<double>(
+                  static_cast<int32_t>(ne.associatedParameterId))));
         o.Set("flags", Napi::Number::New(env, static_cast<double>(ne.flags)));
         return o;
     });
